BookTitle Pagination, Search, ItemsPerPage added, Header and Footer Animation added

+++ views/SEIP137008/BookTitle/trash-item.php
@@ -7,6 +7,26 @@ use App\Utility\Utility;
 $objbooktitle = new BookTitle();
 $allData = $objbooktitle->trash_item('obj');
 
+######################## pagination code block# 1 of 2 start ######################################
+$recordCount= count($allData);
+
+
+if(isset($_REQUEST['Page']))   $page = $_REQUEST['Page'];
+else if(isset($_SESSION['Page']))   $page = $_SESSION['Page'];
+else   $page = 1;
+$_SESSION['Page']= $page;
+
+if(isset($_REQUEST['ItemsPerPage']))   $itemsPerPage = $_REQUEST['ItemsPerPage'];
+else if(isset($_SESSION['ItemsPerPage']))   $itemsPerPage = $_SESSION['ItemsPerPage'];
+else   $itemsPerPage = 5;
+$_SESSION['ItemsPerPage']= $itemsPerPage;
+
+$pages = ceil($recordCount/$itemsPerPage);
+$someData = $objbooktitle->indexPaginator($page,$itemsPerPage);
+
+$serial = (($page-1) * $itemsPerPage) +1;
+
+####################### pagination code block# 1 of 2 end #########################################
 
 ?>
 <!DOCTYPE html>
@@ -22,7 +42,9 @@ $allData = $objbooktitle->trash_item('obj');
     <!-- CSS -->
     <link rel="stylesheet" href="http://fonts.googleapis.com/css?family=Roboto:400,100,300,500">
     <link rel="stylesheet" href="../../../resource/assets/bootstrap/css/bootstrap.min.css">
+    <link rel="stylesheet" href="../../../resource/assets/bootstrap/text-animation/animate.css">
     <link rel="stylesheet" href="../../../resource/assets/font-awesome/css/font-awesome.min.css">
+    <link rel="stylesheet" href="//code.jquery.com/ui/1.12.1/themes/base/jquery-ui.css">
     <link href="https://fonts.googleapis.com/css?family=Lobster|Merriweather|Montserrat|Shrikhand" rel="stylesheet">
     <link rel="stylesheet" href="../../../resource/assets/bootstrap/css/atomic-style.css">
 
@@ -45,8 +67,8 @@ $allData = $objbooktitle->trash_item('obj');
                 <img src="../../../resource/assets/img/logo/book-title.png" alt="..." class="img-rounded">
             </div>
             <div class="header-info col-md-4">
-                <h2 class="project-heading">Atomic Project</h2>
-                <h3 class="project-sub-heading">SEIP 137008 B35 Web Application PHP</h3>
+                <h2 class="project-heading animate-head-text" >Atomic Project</h2>
+                <h3 class="project-sub-heading animate-head-text">SEIP 137008 B35 Web Application PHP</h3>
             </div>
             <div class="user-img .col-md-4">
                 <img src="../../../resource/assets/img/user-img.jpg" alt="..."  class="img-circle size">
@@ -83,9 +105,11 @@ $allData = $objbooktitle->trash_item('obj');
                                  <span class="control-link">Control</span><a href="#" id="button-navbar" class="control-link"><span class="glyphicon glyphicon-chevron-right" ></span></a>
                                     </div>
                                     <div class="nav navbar-nav" role="group" aria-label="..." id="navbar-ctrl">
-                                        <a href="create.php" class="navbar-btn btn btn-info" >Add Item</a>
-                                        <a href="index.php" class="navbar-btn btn btn-warning">Index</a>
-                                        <a href="" class="navbar-btn btn btn-success">PDF Download</a>
+                                        <a href="create.php" class="navbar-btn btn btn-info" ><span class="glyphicon glyphicon-plus-sign"></span> Add Item</a>
+                                        <a href="index.php" class="navbar-btn btn btn-warning"><span class="glyphicon glyphicon-plus-sign"></span> Index</a>
+                                        <a href="pdf.php" class="navbar-btn btn btn-success ">PDF</a>
+                                        <a href="xl.php" class="navbar-btn btn btn-success ">Excel</a>
+                                        <a href="email.php" class="navbar-btn btn btn-success"><span class="glyphicon glyphicon-envelope"></span> Email</a>
                                         <a href="" class="navbar-btn btn btn-primary">Log out</a>
                                     </div>
                                 </div>
@@ -103,8 +127,6 @@ $allData = $objbooktitle->trash_item('obj');
                                 <th colspan="4">Action</th>
                             </tr>
                             <?php
-
-                            $serial = 1;
                             foreach($allData as $oneData){
                                 echo "<tr>";
                                 echo "<td>$serial</td>";
@@ -112,16 +134,16 @@ $allData = $objbooktitle->trash_item('obj');
                                 echo "<td>$oneData->book_name</td>";
                                 echo "<td>$oneData->author_name</td>";
                                 echo " <td>
-                                            <a href='view.php?id=$oneData->id' class='btn btn-info btn-sm'>View</a>
+                                            <a href='view.php?id=$oneData->id' class='btn btn-info btn-sm'><span class='glyphicon glyphicon-eye-open'></span> View</a>
                                             </td>
                                             <td>
-                                            <a href='edit.php?id=$oneData->id' class='btn btn-primary btn-sm'>Edit</a>
+                                            <a href='edit.php?id=$oneData->id' class='btn btn-primary btn-sm'><span class='glyphicon glyphicon-pencil'></span> Edit</a>
                                             </td>
                                             <td>
                                             <a href='clear_trash.php?id=$oneData->id' class='btn btn-success btn-sm'>Reset</a>
                                             </td>
                                             <td>
-                                            <a href='delete.php?id=$oneData->id' class='btn btn-danger btn-sm'>Delete</a>
+                                            <a href='delete.php?id=$oneData->id' class='btn btn-danger btn-sm' onclick='return confirm_msg();'><span class='glyphicon glyphicon-remove'></span> Delete</a>
                                             </td>";
                                 echo "</tr>";
                                 $serial++;
@@ -129,16 +151,46 @@ $allData = $objbooktitle->trash_item('obj');
 
 
                             ?>
-                            <tr class="info">
-                                <td  colspan="8">PAGE:
-                                    <a href="#"><</a>
-                                    <a href="#">1</a>
-                                    <a href="#">2</a>
-                                    <a href="#">3</a>
-                                    <a href="#">4</a>
-                                    <a href="#">5</a>
-                                    <a href="#">6</a>
-                                    <a href="#">></a>
+                            <tr class="">
+                                <td colspan="2">
+                                    <select  class="form-control"  name="ItemsPerPage" id="ItemsPerPage" onchange="javascript:location.href = this.value;" >
+                                        <?php
+                                        if($itemsPerPage==5 ) echo '<option value="?ItemsPerPage=5" selected >Show 5 Items Per Page</option>';
+                                        else echo '<option  value="?ItemsPerPage=5">Show 5 Items Per Page</option>';
+
+                                        if($itemsPerPage==10 )  echo '<option  value="?ItemsPerPage=10" selected >Show 10 Items Per Page</option>';
+                                        else  echo '<option  value="?ItemsPerPage=10">Show 10 Items Per Page</option>';
+
+                                        if($itemsPerPage==15 )  echo '<option  value="?ItemsPerPage=15" selected >Show 15 Items Per Page</option>';
+                                        else echo '<option  value="?ItemsPerPage=15">Show 15 Items Per Page</option>';
+
+                                        if($itemsPerPage==20 )  echo '<option  value="?ItemsPerPage=20"selected >Show 20 Items Per Page</option>';
+                                        else echo '<option  value="?ItemsPerPage=20">Show 20 Items Per Page</option>';
+
+                                        ?>
+                                    </select>
+
+                                </td>
+                                <td  colspan="6">
+                                    <ul class="pagination">
+                                        <?php
+
+                                        $pageMinusOne  = $page-1;
+                                        $pagePlusOne  = $page+1;
+                                        if($page>$pages) Utility::redirect("index.php?Page=$pages");
+
+                                        if($page>1)  echo "<li><a href='index.php?Page=$pageMinusOne'>" . "Previous" . "</a></li>";
+                                        for($i=1;$i<=$pages;$i++)
+                                        {
+                                            if($i==$page) echo '<li class="active"><a href="">'. $i . '</a></li>';
+                                            else  echo "<li><a href='?Page=$i'>". $i . '</a></li>';
+
+                                        }
+                                        if($page<$pages) echo "<li><a href='index.php?Page=$pagePlusOne'>" . "Next" . "</a></li>";
+
+                                        ?>
+
+                                    </ul>
                                 </td>
                             </tr>
 
@@ -161,14 +213,14 @@ $allData = $objbooktitle->trash_item('obj');
 <div class="footer">
     <div class="row">
         <div class="header-info col-md-6">
-            <h4>Atomic Project</h4>
-            <h5> &copy; 2016 Reserved By Efthaqur Alam</h5>
+            <h4 class="animate-footer-text">Atomic Project</h4>
+            <h5 class="animate-footer-text"> &copy; 2016 Reserved By Efthaqur Alam</h5>
         </div>
         <div class="user-img col-md-6">
         </div>
     </div>
 </div>
-</div>
+
 
 <!-- Javascript -->
 <script src="../../../resource/assets/js/jquery-1.11.1.min.js"></script>
@@ -183,6 +235,21 @@ $allData = $objbooktitle->trash_item('obj');
 
 </script>
 
+<script src="../../../resource/assets/bootstrap/text-animation/jquery.fittext.js"></script>
+<script src="../../../resource/assets/bootstrap/text-animation/jquery.lettering.js"></script>
+<script src="../../../resource/assets/bootstrap/text-animation/jquery.textillate.js"></script>
+<script>
+    $('.animate-head-text').textillate({
+        in: { effect: 'wobble' },
+        out: { effect: 'rollOut', sequence: true },
+        loop: true
+    });
+    $('.animate-footer-text').textillate({
+        in: { effect: 'bounceIn' },
+        out: { effect: 'flash', sequence: true },
+        loop: true
+    });
+</script>
 
 </body>
